Add files via upload
The Law of Time 441 Time Cube / #Blockchain cube as the cosmic key to the cosmic cube that is the source geometric shape for all material form.  The Mayan 13 Moon calendar is more accurate than our calendar today: "A clock does not measure time. "A clock measures increments of space which, projected as increments of time, are valued (valorized) into monetary units" // "The solution we propose begins with a time - stamp server" Satoshi Nakamoto // The Law of Time 441 Time Cube http://sawconcepts.com/index/id63.html