Add files via upload
Harvesting Gold: Thomas Edison?s Experiment to Re-Invent American Money
David L. Hammes, Harvesting Gold: Thomas Edison?s Experiment to Re-Invent American Money. Silver City, NM: Richard Mahler Publications, 2012. vi +154 pp. $13 (paperback), ISBN: 978-0-9850667-03. hile he disliked the gold standard he was not willing to give it up in order to limit disruptions and to make his plan more acceptable. The goal was to circumvent its deflationary tendencies, while also preventing inflation and the boom and bust cycles that prevailed until the early part of the twentieth century. At the same time, Edison was also worried about farmers who were heavily indebted but saw the price of their crops decline dramatically. This led him to develop a monetary framework in which gold and agricultural commodities play the role of anchor for the monetary system.

In his system, government would create numerous warehouses for commodities that would buy commodities through a one-year repurchase agreement and the issuance of a warehouse certificate. The inflow of cash to farmers via the repurchase agreement would cover 50 percent of the value of the commodities. Commodities ought to be valued at a price based on a 25-year average instead of current market price. The repurchase agreement must be repaid in 12 monthly installments (i.e. farmers are required to repurchase some of their commodities every month over a twelve-month period at the warehouse price). The certificates would be tradable or could be pledged for a loan, and anybody with a certificate could go to the warehouse and claim ownership of some commodities that would be purchased with 50 percent cash and the handing of the certificate.